Integrate CQRS fixes after rebase (#118)
Retain main's canonical observation construction while restoring the CQRS
helpers and their required context and logging dependencies. Share command
and event fixtures between CQRS tests and record the new fixture module in
the native-wheel snapshot.
